Emails: Extract Google Workspace price component from Email Comparison page
