+++ client/my-sites/email/email-providers-comparison/price/style.scss
@@ -1,3 +1,7 @@
+google-workspace-price__discount {
+	margin-bottom: 12px;
+}
+
 professional-email-price__discount.badge {
 	margin-bottom: 12px;
 }
